backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
twilio==8.10.0
sendgrid==6.10.0
zendesk==1.1.1
python-dotenv==1.0.0
openai==1.0.0
redis==5.0.0
sqlalchemy==2.0.0
pydantic==2.0.0
python-jose==3.3.0
requests==2.31.0
